requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
requests==2.32.4
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
